+++ RuoYi-Vue-v3.8.8/ruoyi-system/src/main/java/com/ruoyi/system/mapper/LocationMapper.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface LocationMapper {
     public List<Location> getProvinceList();
-    public List<Location> getCityList(Integer provinceCode);
-    public List<Location> getAreaList(Integer cityCode);
+    public List<Location> getCityList(Long provinceCode);
+    public List<Location> getAreaList(Long cityCode);
 
 }
